Fix a bug: error when edit or add a GmoStrain without plasmid
In plasmid collection fields, StrainPlasmidType State field is a select list with 2 possibilities, and on per default. But if you don't want to add a plasmid and dont delete the line, you have a bug because on field is not empty (State field) and the user don't have selected a plasmid.

For this bug: use a placeholder: -- select a state -- in state field.