Team Add/Edit: Debug auto addMembers when admins or modos are sets
With PrePersist/PreUpdate, it's work fine just with prePersist, because it's a new object with nothing in database.
But preUpdate don't work fine, because Doctrine can't interract on ManyToMany relation before flush, because it doesn't have informations on the database.

That's why, we use here onFlush, at this state, we have all informations, and we can edit the relation.